requirements.txt.txt
Django==5.2.12
requests==2.32.5
gunicorn==23.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
django-cors-headers==4.7.0
reportlab==4.4.1
